refactor(edges): unify isEdge overlay routing with classic edge engine
- extract roundManhattanCorners (R=5 quadratic fillet) from
  GraphDataElements into edges/routing/classic/round.ts; LVoidEdge
  delegates to the shared export
- EdgeOverlay Manhattan branch now routes via the classic router's
  chooseManhattanSidesAndWaypoints + the shared fillet, replacing
  chooseSides/sideMidpoint/buildPathFromSides and the editor-v2
  roundManhattanPath import; first/last leg clipped to node borders
  overlay-side (revived clipToRect)
- behavior: for center-anchored isEdge arcs the path is geometrically
  the same Z as before (verified: identical turn points); the visible
  net change is corner radius 8 -> 5, now matching native, plus a single
  shared routing implementation. Arrowheads remain Phase 2C.
- routing stays inside memoized EdgeRenderItem; selector untouched
