+++ frontend/src/edges/routing/classic/index.ts
@@ -1,6 +1,7 @@
 export { computeRouting } from './segments';
 export { computeHeadPosition } from './markers';
 export { svgLetterSize } from './stride';
+export { roundManhattanCorners } from './round';
 export type {
     HeadPosition,
     RoutingInput,
